fixed some problems with index

+++ SICS/src/type/PatternMatrix.cpp
@@ -19,7 +19,7 @@ int PatternMatrix::countItems() const {
 
 int PatternMatrix::countIndividuals() const {
 
-	map<boost::dynamic_bitset<>, long int>::const_iterator it;
+	map<boost::dynamic_bitset<>, int>::const_iterator it;
 	int counter = 0;
 
 	for (it = matrix.begin(); it != matrix.end(); ++it) {
@@ -45,7 +45,7 @@ void PatternMatrix::flush() {
 	matrix.clear();
 }
 
-long int & PatternMatrix::operator()(boost::dynamic_bitset<> n) {
+int & PatternMatrix::operator()(boost::dynamic_bitset<> n) {
 	return (matrix[n]);
 }
